Python TwitterAPI to flash Arduino Light
This program flashes a light on an Arduino board via Wifi (same LAN) using the TwitterAPI to watch Twitter feeds. Based on either a tweet containing a specific hashtag or a tweet by a specific handle (configurable), a UDP packet will be sent over Wifi to the Arduino board to flash a light for a configurable amount of time.